LayoutTests/inspector/styles/undo-add-property-expected.txt

Tests that adding a property is undone properly.

Initial value
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
#container  { (undo-add-property.html:7)
font-weight: bold;

[expanded] 
div  { (user agent stylesheet)
display: block;


After adding property
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
#container  { (undo-add-property.html:7)
font-weight: bold;
margin-left: 1px;

[expanded] 
div  { (user agent stylesheet)
display: block;


After undo
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
#container  { (undo-add-property.html:7)
font-weight: bold;

[expanded] 
div  { (user agent stylesheet)
display: block;
